[VL] Follow up on https://github.com/apache/incubator-gluten/pull/9384 to avoid swallow exceptions in UT
